+++ omnigibson/prims/rigid_prim.py
@@ -334,8 +334,8 @@ class RigidPrim(XFormPrim):
         orientation = current_orientation if orientation is None else orientation
 
         # Convert to th.Tensor if necessary
-        position = th.as_tensor(position)
-        orientation = th.as_tensor(orientation)
+        position = th.as_tensor(position, dtype=th.float32)
+        orientation = th.as_tensor(orientation, dtype=th.float32)
 
         # Convert to from scene-relative to world if necessary
         if frame == "scene":
